# perf: more efficient data/v1 POST handler